misc: rename some files in `src/ui/settings/`
