+++ src/system/settings/settings-gamepad.ts
@@ -2,7 +2,7 @@ import { globalScene } from "#app/global-scene";
 import { Button } from "#enums/buttons";
 import { UiMode } from "#enums/ui-mode";
 import { SettingKeyboard } from "#system/settings-keyboard";
-import type { SettingsGamepadUiHandler } from "#ui/settings-gamepad-ui-handler";
+import type { SettingsGamepadUiHandler } from "#ui/gamepad-settings-ui-handler";
 import { truncateString } from "#utils/common";
 import i18next from "i18next";
