+++ src/system/settings/settings-language.ts
@@ -1,6 +1,6 @@
 import { globalScene } from "#app/global-scene";
+import type { SettingsDisplayUiHandler } from "#ui/display-settings-ui-handler";
 import type { LoginRegisterInfoContainerUiHandler } from "#ui/login-register-info-container-ui-handler";
-import type { SettingsDisplayUiHandler } from "#ui/settings-display-ui-handler";
 import i18next from "i18next";
 
 const cancelHandler = () => {
